# backend/src/main/resources/openapi/switchboard-api.yaml
openapi: 3.0.3
info:
  title: Switchboard API
  description: AI-powered feature flag platform. Management surface (user bearer auth), SDK surface (SDK-key auth), and job triggers.
  version: 0.1.0
servers:
  - url: http://localhost:28080
security:
  - bearerAuth: []
...
paths:
...
  /api/projects/{projectId}/ai/proposals:
    post:
      tags: [Ai]
      operationId: draftProposal
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalDraftRequest'
      responses:
        '201':
          description: Drafted proposal with typed diff and rationale
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiProposalResponse'
        '503':
          description: AI unavailable (no API key configured)
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
    get:
      tags: [Ai]
      operationId: listProposals
      parameters:
        - $ref: '#/components/parameters/ProjectId'
        - name: status
          in: query
          schema:
            type: string
            enum: [DRAFT, APPLIED, REJECTED, EXPIRED]
        - name: cursor
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            default: 25
      responses:
        '200':
          description: Proposals, newest first
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiProposalListResponse'
  /api/ai/proposals/{proposalId}:
    get:
      tags: [Ai]
      operationId: getProposal
      parameters:
        - $ref: '#/components/parameters/ProposalId'
      responses:
        '200':
          description: Proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiProposalResponse'
  /api/ai/proposals/{proposalId}/apply:
    post:
      tags: [Ai]
      operationId: applyProposal
      parameters:
        - $ref: '#/components/parameters/ProposalId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalActionRequest'
      summary: Apply a DRAFT proposal, subject to the environment's approval policy.
      description: >
        The write an AI proposal makes is a flag write, so it answers to the same
        per-environment approval policy a hand edit does. In an environment with
        requireApproval on, applying a proposal WRITES NOTHING and opens a PENDING
        change request per environment instead, stamped with this proposal's id; the
        proposal stays DRAFT and becomes APPLIED when the last of those requests is
        approved and applied. A declined or stale request leaves the proposal DRAFT,
        so it can simply be applied again.

        THE ONE EXCEPTION IS AUTOMATION. The rollout monitor's auto-rollback and
        auto-optimize applies bypass this gate while the environment keeps
        allowAutomationBypass on, which is the default - a healing rollback that
        waits for a reviewer during an error spike is not healing. Those writes are
        audited as usual and additionally recorded as APPROVAL_BYPASS. This endpoint
        is a human path, so it never takes the bypass.
      responses:
        '200':
          description: Applied proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiProposalResponse'
        '202':
          description: >
            The environment requires approval. NOTHING WAS WRITTEN: the flag is
            unchanged, the proposal is still DRAFT, and the body is the PENDING
            change request that now stands in for this apply. Clients must branch on
            the status code, not the body. The Location header points at the change
            request. When the proposal touched more than one environment, one request
            was opened per environment and this is the first of them.
          headers:
            Location:
              description: The change request this apply was parked as.
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChangeRequestResponse'
        '409':
          description: Proposal is not in DRAFT status, or already has an open change request
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiError'
  /api/ai/proposals/{proposalId}/reject:
    post:
      tags: [Ai]
      operationId: rejectProposal
      parameters:
        - $ref: '#/components/parameters/ProposalId'
      requestBody:
        required: false
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProposalActionRequest'
      responses:
        '200':
          description: Rejected proposal
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AiProposalResponse'
  /api/environments/{envId}/anomalies:
    get:
      tags: [Ai]
      operationId: listAnomalies
      parameters:
        - $ref: '#/components/parameters/EnvId'
        - name: status
          in: query
          schema:
            type: string
            enum: [OPEN, ACKED, AUTO_ROLLED_BACK]
      responses:
        '200':
          description: Anomaly findings, newest first
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnomalyFindingResponse'
  /api/anomalies/{anomalyId}/ack:
    post:
      tags: [Ai]
      operationId: ackAnomaly
      parameters:
        - name: anomalyId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Acknowledged finding
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnomalyFindingResponse'
  /api/environments/{envId}/flags/{flagKey}/rollout-stats:
    get:
      tags: [Ai]
      operationId: getRolloutStats
      parameters:
        - $ref: '#/components/parameters/EnvId'
        - $ref: '#/components/parameters/FlagKey'
        - name: hours
          in: query
          schema:
            type: integer
            default: 48
      responses:
        '200':
          description: Per-variant eval counts and metric rates, time-bucketed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RolloutStatsResponse'
  /api/users/me/permissions:
    get:
      tags: [Access]
      operationId: getMyPermissions
      summary: What the caller may do at one scope, for hiding controls they cannot use.
      description: >
        Effective permissions are the UNION of the roles the caller holds at the
        requested scope and at every scope containing it. Pass envId for the
        narrowest answer, projectId for the project, or orgId alone. Exactly one of
        the three should be given; the narrowest one wins if more than one is.
      parameters:
        - name: orgId
          in: query
          schema:
            type: string
            format: uuid
        - name: projectId
          in: query
          schema:
            type: string
            format: uuid
        - name: envId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: The caller's effective permissions at that scope
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MyPermissionsResponse'
...
  /api/orgs/{orgId}/role-assignments:
    get:
      tags: [Access]
      operationId: listRoleAssignments
      summary: Role assignments anywhere under this org. Requires MANAGE_MEMBERS.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: scopeType
          in: query
          schema:
            $ref: '#/components/schemas/ScopeType'
        - name: scopeId
          in: query
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Assignments
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignmentListResponse'
    post:
      tags: [Access]
      operationId: grantRole
      summary: Grant a role at a scope. Requires MANAGE_MEMBERS on the org.
      description: >
        Upserts on (user, scope): granting again at the same scope replaces the
        role held there. The scope must belong to this org.
      parameters:
        - $ref: '#/components/parameters/OrgId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RoleAssignmentCreateRequest'
      responses:
        '201':
          description: The assignment
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RoleAssignmentResponse'
  /api/orgs/{orgId}/role-assignments/{assignmentId}:
    delete:
      tags: [Access]
      operationId: revokeRole
      summary: Revoke one role assignment. Requires MANAGE_MEMBERS on the org.
      parameters:
        - $ref: '#/components/parameters/OrgId'
        - name: assignmentId
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Revoked
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: Firebase ID token, or dev token (Bearer dev:email) in local profile.
    sdkKeyAuth:
      type: http
      scheme: bearer
      description: Environment SDK key (sb_srv_...).
    ofrepApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: >
        OFREP's apiKey scheme, carrying the same environment SDK key (sb_srv_...) as sdkKeyAuth.
        Both headers resolve to the same SdkKeyPrincipal; Authorization wins when both are sent.
  parameters:
    OrgId:
      name: orgId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    ProjectId:
      name: projectId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    EnvId:
      name: envId
      in: path
      required: true
      schema:
        type: string
        format: uuid
    FlagKey:
      name: flagKey
      in: path
      required: true
      schema:
        type: string
...
    ProposalId:
      name: proposalId
      in: path
      required: true
      schema:
        type: string
        format: uuid
...
  schemas:
    ApiError:
      type: object
      required: [error, message]
      properties:
        error:
          type: string
          enum: [NOT_FOUND, FORBIDDEN, CONFLICT, VALIDATION_FAILED, UNAUTHORIZED, AI_UNAVAILABLE, UPGRADE_REQUIRED]
        message:
          type: string
...
    FlagKind:
      type: string
      enum: [BOOLEAN, STRING]
...
    VariationCreate:
      type: object
      required: [value]
      properties:
        value:
          type: string
        name:
          type: string
...
    FlagTargetingConfig:
      type: object
      required: [offVariationId, defaultVariationId, fallthrough]
      properties:
        individualTargets:
          type: array
          items:
            $ref: '#/components/schemas/IndividualTarget'
        rules:
          type: array
          items:
            $ref: '#/components/schemas/Rule'
        fallthrough:
          $ref: '#/components/schemas/RolloutOrVariation'
        offVariationId:
          type: string
          format: uuid
        defaultVariationId:
          type: string
          format: uuid
...
    ProposalDraftRequest:
      type: object
      required: [prompt]
      properties:
        prompt:
          type: string
          minLength: 1
          maxLength: 2000
        environmentKey:
          type: string
        flagKey:
          type: string
          description: Set when the prompt edits an existing flag.
    ProposalActionRequest:
      type: object
      properties:
        reason:
          type: string
    ProposalKind:
      type: string
      enum: [FLAG_CREATE, FLAG_UPDATE, ROLLBACK, RETIREMENT]
    ProposalStatus:
      type: string
      enum: [DRAFT, APPLIED, REJECTED, EXPIRED]
    EnvChange:
      type: object
      required: [envKey]
      properties:
        envKey:
          type: string
        enabled:
          type: boolean
        killSwitchActive:
          type: boolean
        config:
          $ref: '#/components/schemas/FlagTargetingConfig'
    FlagChangeDiff:
      type: object
      required: [kind, flagKey]
      properties:
        kind:
          $ref: '#/components/schemas/ProposalKind'
        flagKey:
          type: string
        name:
          type: string
        description:
          type: string
        flagKind:
          $ref: '#/components/schemas/FlagKind'
        variations:
          type: array
          items:
            $ref: '#/components/schemas/VariationCreate'
        tags:
          type: array
          items:
            type: string
        envChanges:
          type: array
          items:
            $ref: '#/components/schemas/EnvChange'
        rollbackToVersion:
          type: integer
        retirementChecklist:
          type: array
          items:
            type: string
    AiProposalResponse:
      type: object
      required: [id, orgId, projectId, kind, diff, status, createdBy, createdAt]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        kind:
          $ref: '#/components/schemas/ProposalKind'
        sourcePrompt:
          type: string
        diff:
          $ref: '#/components/schemas/FlagChangeDiff'
        rationale:
          type: string
        status:
          $ref: '#/components/schemas/ProposalStatus'
        createdBy:
          type: string
        appliedBy:
          type: string
        appliedVersion:
          type: integer
        createdAt:
          type: string
          format: date-time
    AiProposalListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/AiProposalResponse'
        nextCursor:
          type: string
    AnomalyStatus:
      type: string
      enum: [OPEN, ACKED, AUTO_ROLLED_BACK]
    AnomalyFindingResponse:
      type: object
      required: [id, environmentId, flagKey, metricKey, baselineRate, variantRate, zScore, status, createdAt]
      properties:
        id:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        flagKey:
          type: string
        variationId:
          type: string
          format: uuid
        metricKey:
          type: string
        baselineRate:
          type: number
        variantRate:
          type: number
        zScore:
          type: number
        summary:
          type: string
        status:
          $ref: '#/components/schemas/AnomalyStatus'
        suggestedProposalId:
          type: string
          format: uuid
        createdAt:
          type: string
          format: date-time
    VariantStats:
      type: object
      required: [variationId, evalCount, errorRate, conversionRate]
      properties:
        variationId:
          type: string
          format: uuid
        variationName:
          type: string
        evalCount:
          type: integer
          format: int64
        errorRate:
          type: number
        conversionRate:
          type: number
    RolloutStatsBucket:
      type: object
      required: [bucketStart, variants]
      properties:
        bucketStart:
          type: string
          format: date-time
        variants:
          type: array
          items:
            $ref: '#/components/schemas/VariantStats'
    RolloutStatsResponse:
      type: object
      required: [flagKey, environmentId, totals, buckets]
      properties:
        flagKey:
          type: string
        environmentId:
          type: string
          format: uuid
        totals:
          type: array
          items:
            $ref: '#/components/schemas/VariantStats'
        buckets:
          type: array
          items:
            $ref: '#/components/schemas/RolloutStatsBucket'
...
    Permission:
      type: string
      description: A capability the backend checks by name. Roles bundle these.
      enum: [FLAG_READ, FLAG_WRITE, FLAG_KILL, FLAG_ROLLBACK, SEGMENT_WRITE, APPROVE_CHANGES, MANAGE_MEMBERS, MANAGE_SDK_KEYS, MANAGE_SETTINGS, MANAGE_PROJECTS, MANAGE_ENVIRONMENTS, VIEW_AUDIT]
    ScopeType:
      type: string
      description: Where a role is granted. Permissions union across containing scopes.
      enum: [ORG, PROJECT, ENVIRONMENT]
...
    RoleAssignmentResponse:
      type: object
      required: [id, userId, userEmail, scopeType, scopeId, roleKey, createdAt, createdBy]
      properties:
        id:
          type: string
          format: uuid
        userId:
          type: string
          format: uuid
        userEmail:
          type: string
        scopeType:
          $ref: '#/components/schemas/ScopeType'
        scopeId:
          type: string
          format: uuid
        roleKey:
          type: string
        createdAt:
          type: string
          format: date-time
        createdBy:
          type: string
    RoleAssignmentListResponse:
      type: object
      required: [items]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/RoleAssignmentResponse'
    RoleAssignmentCreateRequest:
      type: object
      required: [scopeType, scopeId, roleKey]
      description: Identify the user by id or by email; one of the two is required.
      properties:
        userId:
          type: string
          format: uuid
        email:
          type: string
        scopeType:
          $ref: '#/components/schemas/ScopeType'
        scopeId:
          type: string
          format: uuid
        roleKey:
          type: string
    MyPermissionsResponse:
      type: object
      required: [scopeType, scopeId, permissions]
      properties:
        scopeType:
          $ref: '#/components/schemas/ScopeType'
        scopeId:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        permissions:
          type: array
          items:
            $ref: '#/components/schemas/Permission'
...
    ChangeRequestKind:
      type: string
      enum: [TARGETING_UPDATE, KILL_SWITCH, ROLLBACK]
    ChangeRequestStatus:
      type: string
      description: >
        PENDING is the only reviewable state. STALE means the flag moved past the
        request's baseVersion before it could be applied, so it was refused rather
        than allowed to clobber the newer version.
      enum: [PENDING, APPROVED, DECLINED, APPLIED, WITHDRAWN, STALE]
    ReviewDecision:
      type: string
      enum: [APPROVE, DECLINE]
    ChangeRequestPayload:
      type: object
      description: >
        The proposed write, in the shape the flag write endpoints take. Exactly the
        fields of one kind are populated - enabled+config for TARGETING_UPDATE,
        active for KILL_SWITCH, toVersion for ROLLBACK.
      properties:
        enabled:
          type: boolean
        config:
          $ref: '#/components/schemas/FlagTargetingConfig'
        active:
          type: boolean
        toVersion:
          type: integer
    ChangeRequestReviewResponse:
      type: object
      required: [id, reviewerUserId, reviewer, decision, createdAt]
      properties:
        id:
          type: string
          format: uuid
        reviewerUserId:
          type: string
          format: uuid
        reviewer:
          type: string
        decision:
          $ref: '#/components/schemas/ReviewDecision'
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time
    ChangeRequestResponse:
      type: object
      required: [id, orgId, projectId, environmentId, envKey, flagId, flagKey, kind, payload, baseVersion, minApprovals, allowSelfApproval, status, requestedBy, requestedByUserId, createdAt, reviews, approvalsMet]
      properties:
        id:
          type: string
          format: uuid
        orgId:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        environmentId:
          type: string
          format: uuid
        envKey:
          type: string
        flagId:
          type: string
          format: uuid
        flagKey:
          type: string
        kind:
          $ref: '#/components/schemas/ChangeRequestKind'
        payload:
          $ref: '#/components/schemas/ChangeRequestPayload'
        baseVersion:
          type: integer
          description: The head version the author edited against.
        minApprovals:
          type: integer
          description: Snapshotted from the environment when the request was opened.
        allowSelfApproval:
          type: boolean
        status:
          $ref: '#/components/schemas/ChangeRequestStatus'
        requestedByUserId:
          type: string
          format: uuid
        requestedBy:
          type: string
        comment:
          type: string
        createdAt:
          type: string
          format: date-time
        decidedAt:
          type: string
          format: date-time
        appliedVersion:
          type: integer
          description: The config version this request produced once applied.
        aiProposalId:
          type: string
          format: uuid
          description: >
            Set when this request exists because an AI proposal was applied into an
            environment that requires approval. The proposal stays DRAFT until the
            request applies, and becomes APPLIED when the last of its requests lands.
        approvalsMet:
          type: integer
          description: Approvals that count toward the threshold right now.
        reviews:
          type: array
          items:
            $ref: '#/components/schemas/ChangeRequestReviewResponse'
...